backend/requirements_backup.txt
flask
flask-socketio
websockets
redis
google-genai
prisma
pydantic
pydantic-settings
python-dotenv
eventlet


